Add resolvers directory for nexus types/resolvers.
