+++ src/resolvers/index.ts
@@ -1,0 +1,3 @@
+import Query from './Query';
+
+export default { Query };
